Clone the options object in authorizationParams
Function authorizationParams returns extra Azure AD-sepecific parameters
to be included in the authorization request. Here is simply returns the
options object that is passed but the caller then modifies the object
which can cause problems in subsequent authentication attempts. With
this change, a clone of the options object is returned, so the
modifications don't affect subsequent authentications.